requirements.txt
streamlit
pandas
PyJWT